backend/pyproject.toml
[tool.poetry]
name = "aurix-backend"
version = "0.1.0"
description = "Aurix AI-Powered Financial Intelligence Platform - Backend"
authors = ["Aurix Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlmodel = "^0.0.14"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.0"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
redis = "^5.0.1"
celery = "^5.3.4"
pandas = "^2.1.4"
prophet = "^1.1.5"
openai = "^1.10.0"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
stripe = "^7.9.0"
google-auth = "^2.26.2"
google-auth-oauthlib = "^1.2.0"
google-api-python-client = "^2.111.0"
slack-sdk = "^3.26.2"
sendgrid = "^6.11.0"
cryptography = "^42.0.0"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
ruff = "^0.1.11"
mypy = "^1.8.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
